requirements.txt
Cython==0.25.2
fasttext==0.8.3
gensim==2.1.0
JPype1==0.6.2
numpy==1.12.1
scipy==0.19.0
sklearn-crfsuite==0.3.5
